git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@1017 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Utils/Animated ProgressBar.cpp
@@ -144,7 +144,7 @@ void SetProgressBarTitle( UINT32 ubID, STR16 pString, UINT32 usFont, UINT8 ubFor
 	}
 	if( pString && wcslen( pString ) )
 	{
-		pCurr->swzTitle = (STR16) MemAlloc( sizeof( CHAR16 ) * ( wcslen( pString ) + 1 ) );
+		pCurr->swzTitle = (STR16) MemAlloc( (UINT32) ( sizeof( CHAR16 ) * ( wcslen( pString ) + 1 ) ) );
 		swprintf( pCurr->swzTitle, pString );
 	}
 	pCurr->usTitleFont = (UINT16)usFont;
